# tests/unit/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "ApplicantRead": {
        "properties": {
          "current_ctc": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Ctc"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "notice_period_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notice Period Days"
          },
          "years_experience": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Years Experience"
          }
        },
        "required": [
          "id",
          "full_name",
          "notice_period_days",
          "current_ctc",
          "years_experience"
        ],
        "title": "ApplicantRead",
        "type": "object"
      },
...
      "Body_upload_resume_v1_applicants_me_resumes_post": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_resume_v1_applicants_me_resumes_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MeResponse": {
        "properties": {
          "applicant": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApplicantRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "role"
        ],
        "title": "MeResponse",
        "type": "object"
      },
...
      "ProfileUpdate": {
        "additionalProperties": false,
        "description": "Partial profile update. Only keys present in the request are applied\n(`model_fields_set`); an explicit null clears a nullable column.\n`full_name` is non-nullable and rejects an explicit null.",
        "properties": {
          "current_ctc": {
            "anyOf": [
              {
                "maximum": 9999999999.99,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Ctc"
          },
          "full_name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "notice_period_days": {
            "anyOf": [
              {
                "maximum": 365.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notice Period Days"
          },
          "years_experience": {
            "anyOf": [
              {
                "maximum": 60.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Years Experience"
          }
        },
        "title": "ProfileUpdate",
        "type": "object"
      },
...
      "ResumeParseStatus": {
        "enum": [
          "pending",
          "parsing",
          "parsed",
          "failed"
        ],
        "title": "ResumeParseStatus",
        "type": "string"
      },
      "ResumeRead": {
        "description": "Response shape for resume metadata. Bytes are never returned here.",
        "properties": {
          "applicant_id": {
            "format": "uuid",
            "title": "Applicant Id",
            "type": "string"
          },
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "original_filename": {
            "title": "Original Filename",
            "type": "string"
          },
          "parse_status": {
            "$ref": "#/components/schemas/ResumeParseStatus"
          },
          "parsed_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parsed Json"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "applicant_id",
          "original_filename",
          "content_type",
          "size_bytes",
          "parse_status",
          "created_at"
        ],
        "title": "ResumeRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Jobify API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/applicants/me": {
      "patch": {
        "operationId": "update_profile_v1_applicants_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Profile",
        "tags": [
          "applicants"
        ]
      }
    },
...
    "/v1/applicants/me/resumes": {
      "get": {
        "description": "List the authenticated applicant's resumes, newest first.",
        "operationId": "list_resumes_v1_applicants_me_resumes_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ResumeRead"
                  },
                  "title": "Response List Resumes V1 Applicants Me Resumes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Resumes",
        "tags": [
          "resumes"
        ]
      },
      "post": {
        "operationId": "upload_resume_v1_applicants_me_resumes_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_resume_v1_applicants_me_resumes_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Resume",
        "tags": [
          "resumes"
        ]
      }
    },
    "/v1/applicants/me/resumes/{resume_id}": {
      "get": {
        "operationId": "get_resume_v1_applicants_me_resumes__resume_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "resume_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Resume Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Resume",
        "tags": [
          "resumes"
        ]
      }
    },
...
  }
}